// 1. Logic integration - This code will be the full connector that takes the data provided from the other sources and takes the compilation from main.cpp to direct thruster control ----------------- Added // 2. Thruster logic //      - Thrusters must have smart logic that will force them to automatically determine the amount of thrust/power they need  ----------------------- Added //      - This will also require me to create limits and maximums for thrust    ------------------ Added //      - This also includes failsafes to prevent thruster burnout       ------------------ Added //      - Implement "power-ramping" (thruster won't go from 0% to 100% in 10 ms)   ------------------ Added //      - Tether tension checks for diving to ensure the tension doesn't drag the float underwater or snap the tether.  --------------------- Can't do without proper components // 3. Code the mutex, so thrust isn't accidentally interrupted and causes the hull to halt movement (unless given an exception like an object) // 4. Ensure the code is able to provide data and integrate into other code -------------------------- Added //      - This means everything else will also have their source code and header files tweaked and added upon to supplement the thrusters // 5. Constant updating loop providing data on the current status of the thrusters ----------------- Added //      - I will likely build a website where all this can be viewed to prevent looking at log/console clutter when the program runs // 6. Speed and RPM tracking ------------------ Can't do without proper componenets //      - Will need to check parts list if we have the capability for that // 7. Communication checks       ------------------ Added //      - Are the thrusting recieving commands properly from the obstacle logic and AI? // 8. Manual ovveride control over thrusters if need be ---------- Added //      - Allows us to stop current program commands from the AI if need be //      - Ideally, this won't stop information given by the AI and obstacle logic // 9. Introduce failsafes to prevent data/file corruption ------------ Added // 10. Create  timekeeps to monitor/predict thruster resilience  ---------- Deemed unecessary for now, but will be added in the future if needed //      - Rough idea whether or not to stop the program and go manual or execute ascension // 11. Introduce heartbeat code to prevent prepetual, unstoppable thrust --------- Added
Add ThrusterController module and integrate with main.cpp

Added ThrusterController class for thruster logic, power ramping, manual override, and safety features. Integrated with main.cpp, improved time handling, and fixed logging bug. Updated project files, improved thread safety, cleaned up comments, and fixed formatting in related modules. Added enums and data structures for propulsion modes.

+++ OpticalSensorFirmware/SonarSensorHub.cpp
@@ -5,7 +5,7 @@
 unsigned long getTheTimeinMillis() noexcept {
     static const auto start = std::chrono::steady_clock::now();
     const auto now = std::chrono::steady_clock::now();
-    return std::chrono::duration_cast<std::chrono::milliseconds>(now - start).count();
+    return static_cast<unsigned long>(std::chrono::duration_cast<std::chrono::milliseconds>(now - start).count());
 }
 
 void TriggerPingThing(int pin) noexcept {
